# back/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/offers/{offerId}/steps": {
      "get": {
        "operationId": "InterviewStepsController_list",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewStepDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No offer with this id (`resource-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No offer has the id 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d. It may have been deleted; list offers (GET /offers) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "offer"
                }
              }
            }
          }
        },
        "summary": "List the steps of an offer, in order",
        "tags": [
          "interview-steps"
        ]
      },
      "post": {
        "operationId": "InterviewStepsController_create",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewStepDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewStepDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No offer with this id (`resource-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No offer has the id 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d. It may have been deleted; list offers (GET /offers) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "offer"
                }
              }
            }
          }
        },
        "summary": "Add a step after the existing ones",
        "tags": [
          "interview-steps"
        ]
      }
    },
    "/offers/{offerId}/steps/order": {
      "put": {
        "description": "Replaces the whole order at once. The list must contain every step of the offer exactly once.",
        "operationId": "InterviewStepsController_reorder",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderInterviewStepsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The steps in their new order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewStepDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No offer with this id (`resource-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No offer has the id 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d. It may have been deleted; list offers (GET /offers) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "offer"
                }
              }
            }
          }
        },
        "summary": "Reorder the steps of an offer",
        "tags": [
          "interview-steps"
        ]
      }
    },
    "/offers/{offerId}/steps/{stepId}": {
      "get": {
        "operationId": "InterviewStepsController_get",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "stepId",
            "required": true,
            "in": "path",
            "description": "Interview step id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewStepDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "The offer has no step with this id, or there is no such offer (`resource-not-found`, `resource` says which).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "Offer 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d has no interview step with the id 0199a7a4-3c2e-7b6a-9c1d-00000000000a. List its steps (GET /offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d/steps) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "interview-step"
                }
              }
            }
          }
        },
        "summary": "Get a step",
        "tags": [
          "interview-steps"
        ]
      },
      "patch": {
        "description": "Partial update: absent fields are left unchanged, `null` clears an optional field.",
        "operationId": "InterviewStepsController_update",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "stepId",
            "required": true,
            "in": "path",
            "description": "Interview step id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInterviewStepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewStepDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "The offer has no step with this id, or there is no such offer (`resource-not-found`, `resource` says which).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "Offer 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d has no interview step with the id 0199a7a4-3c2e-7b6a-9c1d-00000000000a. List its steps (GET /offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d/steps) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "interview-step"
                }
              }
            }
          }
        },
        "summary": "Update a step",
        "tags": [
          "interview-steps"
        ]
      },
      "delete": {
        "operationId": "InterviewStepsController_remove",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "stepId",
            "required": true,
            "in": "path",
            "description": "Interview step id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "The offer has no step with this id, or there is no such offer (`resource-not-found`, `resource` says which).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "Offer 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d has no interview step with the id 0199a7a4-3c2e-7b6a-9c1d-00000000000a. List its steps (GET /offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d/steps) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "interview-step"
                }
              }
            }
          }
        },
        "summary": "Delete a step",
        "tags": [
          "interview-steps"
        ]
      }
    }
  },
  "info": {
    "title": "emploi API",
    "description": "Job search history: the job offers the user responded to and the steps of their interview processes.\n\nErrors are RFC 9457 Problem Details (`application/problem+json`): `type` names the kind of problem (`GET` it, e.g. `/problems/validation-error`, for what it means and how to fix the request), `detail` says what happened and what to do, and a `validation-error` lists every invalid value in `errors`, with its location, the rule broken and, when they help, the accepted values.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InvalidParamDto": {
        "type": "object",
        "properties": {
          "in": {
            "type": "string",
            "enum": [
              "body",
              "query",
              "path"
            ]
          },
          "name": {
            "type": "string",
            "description": "JSON Pointer (RFC 6901) into the body, e.g. `/title`, or the query or path parameter name.",
            "example": "/title"
          },
          "code": {
            "type": "string",
            "description": "Rule broken, stable: e.g. `isNotEmpty`, `maxLength`, `isIn`, `isUuid`, `whitelistValidation` (unknown field: remove it).",
            "example": "isNotEmpty"
          },
          "detail": {
            "type": "string",
            "example": "title should not be empty"
          },
          "allowed": {
            "description": "For `isIn`: the accepted values.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxLength": {
            "type": "integer",
            "description": "For `maxLength`: the maximum number of characters."
          }
        },
        "required": [
          "in",
          "name",
          "code",
          "detail"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Kind of problem: `/problems/<kind>`, relative to the API (`GET` it for what it means and how to fix the request), or `about:blank`.",
            "example": "/problems/validation-error"
          },
          "title": {
            "type": "string",
            "description": "Summary of the kind of problem."
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "What happened this time, and what to do about it."
          },
          "instance": {
            "type": "string",
            "description": "Path of the request that failed."
          },
          "errors": {
            "description": "`validation-error`: every invalid value.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParamDto"
            }
          },
          "resource": {
            "type": "string",
            "enum": [
              "offer",
              "interview-step"
            ],
            "description": "`resource-not-found`: the kind of resource not found."
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "InterviewStepStatus": {
        "type": "string",
        "enum": [
          "planned",
          "pending",
          "passed",
          "failed",
          "cancelled"
        ],
        "description": "`planned`: to come; `pending`: done, waiting for the answer; `passed`: successful; `failed`: rejected at this step; `cancelled`: did not take place."
      },
      "InterviewStepDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "offerId": {
            "type": "string",
            "format": "uuid",
            "description": "The offer the step belongs to."
          },
          "title": {
            "type": "string",
            "description": "What the step is, free text. Trimmed.",
            "maxLength": 200,
            "example": "Entretien téléphonique RH"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Notes: who, what was asked, feedback.",
            "maxLength": 20000
          },
          "date": {
            "type": "string",
            "nullable": true,
            "format": "date",
            "description": "Day of the step, `YYYY-MM-DD`; `null` until scheduled.",
            "example": "2026-10-06"
          },
          "status": {
            "description": "`planned`: to come; `pending`: done, waiting for the answer; `passed`: successful; `failed`: rejected at this step; `cancelled`: did not take place.",
            "example": "planned",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewStepStatus"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-01T08:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T09:45:00.000Z"
          }
        },
        "required": [
          "id",
          "offerId",
          "title",
          "description",
          "date",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateInterviewStepDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "What the step is, free text. Trimmed.",
            "maxLength": 200,
            "example": "Entretien téléphonique RH"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Notes: who, what was asked, feedback.",
            "maxLength": 20000
          },
          "date": {
            "type": "string",
            "nullable": true,
            "format": "date",
            "description": "Day of the step, `YYYY-MM-DD`; `null` until scheduled.",
            "example": "2026-10-06"
          },
          "status": {
            "description": "`planned`: to come; `pending`: done, waiting for the answer; `passed`: successful; `failed`: rejected at this step; `cancelled`: did not take place.",
            "example": "planned",
            "default": "planned",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewStepStatus"
              }
            ]
          }
        },
        "required": [
          "title"
        ]
      },
      "ReorderInterviewStepsDto": {
        "type": "object",
        "properties": {
          "stepIds": {
            "description": "Every step id of the offer, exactly once, in the new order. Any other list is rejected with 400.",
            "uniqueItems": true,
            "maxItems": 500,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "stepIds"
        ]
      },
      "UpdateInterviewStepDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "What the step is, free text. Trimmed.",
            "maxLength": 200,
            "example": "Entretien téléphonique RH"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Notes: who, what was asked, feedback.",
            "maxLength": 20000
          },
          "date": {
            "type": "string",
            "nullable": true,
            "format": "date",
            "description": "Day of the step, `YYYY-MM-DD`; `null` until scheduled.",
            "example": "2026-10-06"
          },
          "status": {
            "description": "`planned`: to come; `pending`: done, waiting for the answer; `passed`: successful; `failed`: rejected at this step; `cancelled`: did not take place.",
            "example": "planned",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewStepStatus"
              }
            ]
          }
        }
      }
    }
  }
}